Data Update Protocol
This had been tested on the Mini receiving update of data from control board. The test included the timing, having the Round Robin on, Serial baud rate and so on. This code is the final result where data will be updated continuously.
There are still debug code in there but I left it there so we can do debugging in the future, comment out if you want to.
Note: The baud rate should be calibrated after all is done to get maximum performance. For now, I left it at 9600.