Added Untested Code for Movement, Liftoff, Land, and UltraSensor
Combined the skeleton on possible protocol for movement, liftoff, land and the Ultrasonic code with our main code.
This skeleton from Eric is also filled in with code and some protocol was modified. The ultrasonic code from Doug is included so that it updates each time UART update is called even if UART updates failed.
The Ultrasonic sensor part should be fine, but the movement, liftoff, and land needs testing.
Look at notebook and comments for detailed information, flowchart needs updating because some protocol is changed.
Added some extra comments on my functions for clarity.