Variable and File Change
command and node code LED variable is no longer hard coded but set to a variable which allows easy access. I also seperated the new read_write function into 2 parts where the new one is called V2. Currently the code uses the old one because the hardwares are not up to date.