Moter Handle & Debug Code
Added motor handle stuff, but the code did not work on the copter setup, the debugging code showed that the things get stuck at init of cc2500. Sometimes it gets passed that but it still will not work with sending of packets. So something should be wrong with init and have to ask doug about this.